requirements.txt
google-genai==1.64.0
langchain-community==0.3.31
langchain-text-splitters==0.3.11
pymilvus==2.6.9
pypdf==6.1.3
python-dotenv==1.2.1
tiktoken==0.12.0
